server/requirements.txt
Flask==3.0.3
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
python-dotenv==1.0.1

